requirements.txt
beautifulsoup4==4.12.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
langchain-openai==0.0.2
langchain-core==0.1.8
openai>=1.6.1,<2.0.0
pdf2image==1.16.3
python-dotenv==1.0.0
requests==2.31.0
watchdog==3.0.0
